+++ src/main/java/nextstep/member/UserDetailsFactoryImpl.java
@@ -3,7 +3,7 @@ package nextstep.member;
 import auth.UserDetails;
 import auth.UserDetailsFactory;
 import lombok.RequiredArgsConstructor;
-import nextstep.support.AuthenticationException;
+import nextstep.support.exception.AuthenticationException;
 
 @RequiredArgsConstructor
 public class UserDetailsFactoryImpl implements UserDetailsFactory {
