+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -5,10 +5,10 @@ import nextstep.member.Member;
 import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
-import nextstep.support.AlreadyReservedScheduleException;
-import nextstep.support.AuthenticationException;
-import nextstep.support.DuplicateEntityException;
-import nextstep.support.NoReservationException;
+import nextstep.support.exception.AlreadyReservedScheduleException;
+import nextstep.support.exception.AuthenticationException;
+import nextstep.support.exception.DuplicateEntityException;
+import nextstep.support.exception.NoReservationException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
