+++ src/main/java/nextstep/reservation_waiting/ReservationWaitingService.java
@@ -3,7 +3,7 @@ package nextstep.reservation_waiting;
 import lombok.RequiredArgsConstructor;
 import nextstep.member.Member;
 import nextstep.reservation.Reservation;
-import nextstep.support.DuplicateEntityException;
+import nextstep.support.exception.DuplicateEntityException;
 import org.springframework.stereotype.Service;
 
 @Service
