+++ src/main/java/nextstep/support/exception/AlreadyReservedScheduleException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.support.exception;
 
 public class AlreadyReservedScheduleException extends RuntimeException {
     public AlreadyReservedScheduleException() {
