+++ src/main/java/nextstep/support/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.support.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
